+++ resources/views/default/_partials/navigation/divider.blade.php
@@ -1,0 +1,3 @@
+<li role="separator" id="{{ $id }}">
+    <hr class="border-secondary my-1">
+</li>