+++ resources/views/default/_partials/navigation/label.blade.php
@@ -1,0 +1,3 @@
+<li class="nav-header" id="{{ $id }}">
+    {!! $title !!}
+</li>